Backend/lol-chatter-backend/pyproject.toml
[tool.poetry]
name = "lol-chatter-backend"
version = "0.1.0"
description = ""
authors = ["EzzatEsam <ezzatesam1214@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.111.0"
sqlalchemy = "^2.0.31"
bcrypt = "^4.1.3"
pyjwt = "^2.8.0"
websockets = "^12.0"
google-generativeai = "^0.7.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
